cheri: Manually apply DriverData cast fixes in macros
The clang-tidy check will warn about assignments to driver
data fields with the wrong cast even if the actual assignment
originates from a macro expansion. However, the macro expansion
will not be edited/fixed.

Handle the respective warnings manually.

+++ drivers/acpi/acpi_apd.c
@@ -41,7 +41,7 @@ struct apd_private_data {
 };
 
 #if defined(CONFIG_X86_AMD_PLATFORM_DEVICE) || defined(CONFIG_ARM64)
-#define APD_ADDR(desc)	((unsigned long)&desc)
+#define APD_ADDR(desc)	((uintptr_t)&desc)
 
 static int acpi_apd_setup(struct apd_private_data *pdata)
 {
